requirements.txt
fastapi==0.95.1
grpcio==1.52.0
grpcio-tools==1.51.3
protobuf==4.22.1
pydantic==1.10.5
python-dotenv==1.0.0
numpy~=1.24.3
uvicorn==0.22.0
